+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change log
 
+## [Unreleased]
+
+- Updated `codecov-action` version to 4.5.0.
+
 ## 2.0.1 - 2024-08-09
 
 - Updated `README.md` file.